requirements.txt
anthropic
msal
requests
python-dotenv
flask
PyPDF2
python-docx